create contact.java to introduce inheritance to person.java
